mock-requirements.txt
pytest==4.4.0
pytest-mock==1.10.3
requests-mock==1.5.2